requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
openai==1.3.5
requests==2.31.0
jinja2==3.1.2
python-multipart==0.0.6
apscheduler==3.10.4
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
aiofiles==23.2.1
